+++ traitsui/helper.py
@@ -182,3 +182,14 @@ def compute_column_widths(available_space, requested, min_widths, user_widths):
             available_space -= widths[i]
 
     return widths
+
+
+# Wrapper for TraitPrefixList deprecation
+try:
+    from traits.api import PrefixList
+except ImportError:
+    def PrefixList(list_, default_value=None, **kwargs):
+        from traits.api import Trait, TraitPrefixList
+        if default_value is None:
+            default_value = list_[0]
+        return Trait(default_value, TraitPrefixList(list_), **kwargs)
